% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopMocker-package.R
\docType{package}
\name{OmopMocker-package}
\alias{OmopMocker}
\alias{OmopMocker-package}
\title{OmopMocker: Create an OMOP CDM Mock Object}
\description{
This package provides the tools to create mock omop cdm objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopMocker/}
}

}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
