% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCohortSet.R
\name{generateSyntheticCohortSet}
\alias{generateSyntheticCohortSet}
\title{It creates a mock cdm_reference.
Function to generate synthetic Cohort}
\usage{
generateSyntheticCohortSet(
  cdm,
  tableName = "cohort",
  numberCohorts = 1,
  cohortName = paste0("cohort_", seq_len(numberCohorts)),
  recordPerson = 1,
  cdmVersion = attr(cdm, "cdm_version"),
  seed = 1
)
}
\arguments{
\item{cdm}{cdm object}

\item{tableName}{name of the cohort table}

\item{numberCohorts}{number of different cohort to create}

\item{cohortName}{name of the cohort within the table}

\item{recordPerson}{the expected number of record per person}

\item{cdmVersion}{version number of the cdm}

\item{seed}{random seed}
}
\value{
A cdm reference with the mock tables
}
\description{
It creates a mock cdm_reference.
Function to generate synthetic Cohort
}
