% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCdm.R
\name{mockCdm}
\alias{mockCdm}
\title{It creates a mock cdm_reference.}
\usage{
mockCdm(
  cdmName = "mock cdm",
  cdmVersion = "5.3",
  person = NULL,
  observationPeriod = NULL,
  seed = 1
)
}
\arguments{
\item{cdmName}{Name of the cdm.}

\item{cdmVersion}{version number of the cdm}

\item{person}{person table}

\item{observationPeriod}{observation_period table}

\item{seed}{random seed}
}
\value{
A cdm reference with the mock tables
}
\description{
It creates a mock cdm_reference.
}
